moved ff func in pkg
